new apple and windows
new apple and windows code

+++ jamLib/Object.cpp
@@ -6,7 +6,7 @@ namespace jamLib{
 Object::Object(){}
 Object::~Object(){}
 
-void* Object::operator new(unsigned int size) throw()
+void* Object::operator new(size_t size) throw()
 {
     return malloc(size);
 }
@@ -15,7 +15,7 @@ void Object::operator delete(void* p)
      free(p);
 }
 
-void* Object::operator new[](unsigned int size) throw()
+void* Object::operator new[](size_t size) throw()
 {
     return malloc(size);
 }
